requirements.txt
Django==4.1.7
djangorestframework==3.14.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.3
Pillow
django-cors-headers==3.14.0
drf-extra-fields==3.4.1
django-elasticsearch-dsl>=7.0.0
django-elasticsearch-dsl-drf>=0.22.0